rename test-settings
