+++ test/helpers/default-functions.ts
@@ -2,12 +2,12 @@ import axios, { AxiosResponse } from 'axios';
 import lodash from 'lodash';
 import { Context } from 'mocha';
 import WikibaseApi from 'wdio-wikibase/wikibase.api.js';
-import { TestSettings } from '../types/TestSettings.js';
 import Binding from '../types/binding.js';
 import BotResponse from '../types/bot-response.js';
 import DatabaseConfig from '../types/database-config.js';
 import ExternalChange from '../types/external-change.js';
 import LuaCPUValue from '../types/lua-cpu-value.js';
+import { TestSettings } from '../types/test-settings.js';
 
 export function defaultFunctions(): void {
 	const settings: TestSettings = testEnv.settings;
