+++ test/setup/make-test-settings.ts
@@ -8,7 +8,7 @@ import TestSettings, {
 	TestHooks,
 	TestRunnerSettings,
 	TestSuiteSettings
-} from '../types/TestSettings.js';
+} from '../types/test-settings.js';
 import loadEnvFiles from './load-env-files.js';
 
 export const ONE_DAY_IN_MS = 86400000;
